Added authentication router to ETOS API.
Authentication methods for Simple, LDAP and None added.
Simple = username/password authentication, validate against DB
LDAP = username/password authentication, validate against LDAP
None = No credentials

Authentication methods are required in order to request an API
token. This token is always required in order to communicate
with the API, even when using the "NoAuth" method.
To request a token, send a post request to /token with proper
credentials depending on the backend chosen.

A token will be returned which shall be added to the request
headers in the form of 'Authorization: Bearer TOKEN'
The token will expire after a while and a new token must be
requested at that time.

Note that tests to not work yet and no tests for authentication
have been written.