Fix the problem where a wrong realm is found in www-authenticate
The token cache returned an invalid token for the container registry
and when using that token on the first request we got an incorrect
realm in the www-authenticate header.
If that happens we will now invalidate the cache and do a HEAD
request without a token to get a proper www-authenticate.