Add a traceparent annotation to the TestRun created by v1alpha (#112)
* Add a traceparent annotation to the TestRun created by v1alpha

This traceparent shall be picked up by the testrun_controller and
be added as an environment variable to be used in ETOS to link all
traces together.

* Change from label to annotation

We don't need to select pods using the traceparent since it will
be different between all services. Better to have it as an annotation.