A new logarea service where files can be uploaded to
Removed the old logarea service which is no longer in use.
That service was used for more complex report and artifact downloads
by the ETOS client. The Client now receives reports and artifacts
via SSE so that service is no longer in use.

I re-used the logarea name for this new service which can
be used by ETOS to upload reports and artifacts to as a
log area. This service is not designed to be production quality,
more like a logarea that we can deploy in ETOS test environments
and for users to try out ETOS.